feat: overhead routes

+++ src/Meisy.Domain/Entities/Input.cs
@@ -8,8 +8,9 @@ namespace Meisy.Domain.Entities
         public string Description { get; set; } = string.Empty;
         public decimal Price { get; set; }
         public InputType Type { get; set; }
-        public int Amount { get; set; }
+        public double Amount { get; set; }
         public MeasurementUnit MeasurementUnit { get; set; }
+
         public int CompanyId { get; set; }
         public Company Company { get; set; } = default!;
     }
